Rename MUIDataTable to DataTable and update component class name
